sepolicy:peripheral_manager: Add support for the peripheral state property
Peripheral manager now writes to a property for each subsystem under it's
control, containing the current state of the subsystem. It also publishes
a property which contains the list of subsystems required to be gracefully
shutdown. These properties can be used by clients of peripheral manager and
also by the system server during full device shutdown.

Change-Id: I531bc7d5c965b2990ec97bd6bb7af9dc752221d5