sepolicy : remove system_writes_vendor_properties_violators
As part of treble system /core services are not allowed  to
set vendor property .

if Property defined is part of system image then it can be set
or get by core/system services provided we define the property
as extended_core_property_type. So adding this to the property
that are added by vendor and used by core/system services.

Change-Id: I7ad8bc562be09126c082fc54f52499f5138fea5b