sepolicy: Allow USB HAL to search for battery-supply
Commit 444bf782ab3af ("sepolicy: allow USB HAL access to
battery-supply and power-supply") introduced a change to
allow vendor_hal_usb_qti for accessing power and battery
supply. However, search permissions for
vendor_sysfs_battery_supply is required in order to access
this which is missing. This is giving avc denied error
on search for qcom-battery.

Fix this by giving search permissions to usb hal to access
to search for battery supply.

Change-Id: I3724e141b2082c162c9ea5a0edd3c8f9becaeb16
Signed-off-by: Udipto Goswami <ugoswami@codeaurora.org>